changed render clamp implementation to be integer (#102)

+++ sola-engine/platform/swing/src/main/java/technology/sola/engine/platform/swing/core/Graphics2dRenderer.java
@@ -67,7 +67,7 @@ public class Graphics2dRenderer implements Renderer {
   }
 
   @Override
-  public void setClamp(float x, float y, float width, float height) {
+  public void setClamp(int x, int y, int width, int height) {
     throw new NotYetImplementedException();
   }
